moved assets folder to public folder
