+++ src/components/HomePage.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import BackgroundSlider from './BackgroundSlider';
-import MainContent from './MainContent';
-import SignUp from './SignUp';
-import HowItWorks from './HowItWorks';
-import Testimonials from './Testimonials';
+import BackgroundSlider from './backgroundslider/BackgroundSlider';
+import MainContent from './maincontent/MainContent';
+import SignUp from './login/SignUp';
+import HowItWorks from './howitworks/HowItWorks';
+import Testimonials from './testimonal/Testimonials';
 
 const HomePage = () => {
   return (
